Functions on included tex/bib moved to cacher
