+++ src/providers/completer/commandlib/commandfinder.ts
@@ -272,7 +272,7 @@ export class CommandFinder {
      */
     private whichPackageProvidesCommand(cmdName: string): string {
         if (this.extension.manager.rootFile !== undefined) {
-            for (const file of this.extension.manager.getIncludedTeX()) {
+            for (const file of this.extension.cacher.getIncludedTeX()) {
                 const packages = this.extension.cacher.getCachedContent(file)?.element.package
                 if (packages === undefined) {
                     continue
